Fixed but in typedefs and set up query to render all meals on menu page. Also moved mapped listed to their own component files
