+++ client/src/pages/HomePage.jsx
@@ -1,10 +1,10 @@
 /* eslint-disable no-unused-vars */
 import React from 'react';
+import AllMeals from "../components/AllMeals";
+
 
 export default function HomePage() {
     return (
-        <div className=''>
-            <p>This is the Home page!</p>
-            </div>
+            <AllMeals />
     )
 }
